XXX Crate paths seem broken with default localPatterns
This really shouldn't be a problem.  The Nix error is that it doesn't know how
to unpack the archive, which is completely empty because the nix path ends up
not being created.

I checked the logic in the overlay and believe that the type guard on local
paths is correct, so localPatterns should have no effect.  Something annoying is
going on.